BREAKING: update code to new ext:solr 1.5-rc
This adds compatibility with TYPO3 11 but relies on a release candidate
of EXT:solr

+++ Classes/Domain/Search/ResultSet/SearchResultSet.php
@@ -110,7 +110,7 @@ class SearchResultSet extends \ApacheSolrForTypo3\Solr\Domain\Search\ResultSet\S
         $numberOfResults = $this->getAllResultCount();
 
         if ($numberOfResults - (2 * $resultsPerPage) > $resultOffset) {
-            $uri = $this->searchUriBuilder->getResultPageUri($previousRequest, ceil($numberOfResults / $resultsPerPage));
+            $uri = $this->searchUriBuilder->getResultPageUri($previousRequest, (int)ceil($numberOfResults / $resultsPerPage));
         }
         return $uri;
     }
